extract common component elements into base component
