+++ addon/components/file-dropzone/component.js
@@ -1,9 +1,8 @@
 /* global Blob, Uint8Array */
-import Component from '@ember/component';
+import BaseComponent from '../base-component';
 
-import { inject as service } from '@ember/service';
 import { bind } from '@ember/runloop';
-import { computed, set, get } from '@ember/object';
+import { set, get } from '@ember/object';
 import layout from './template';
 import DataTransfer from '../../system/data-transfer';
 import uuid from '../../system/uuid';
@@ -72,18 +71,9 @@ const dragListener = new DragListener();
   @yield {valid} dropzone.valid
   @yield {Queue} queue
  */
-export default Component.extend({
+export default BaseComponent.extend({
 
   layout,
-
-  /**
-    The name of the queue that files should be
-    added to when they get dropped.
-
-    @argument name
-    @type {string}
-   */
-  name: null,
 
   supported,
 
@@ -112,8 +102,6 @@ export default Component.extend({
     @type {function}
    */
   ondrop: null,
-
-  fileQueue: service(),
 
   /**
     Whether users can upload content
@@ -145,15 +133,6 @@ export default Component.extend({
     @default null
    */
   cursor: null,
-
-  queue: computed('name', {
-    get() {
-      let queueName = get(this, 'name');
-      let queues = get(this, 'fileQueue');
-      return queues.find(queueName) ||
-             queues.create(queueName);
-    }
-  }),
 
   didInsertElement() {
     this._super();
